fix 001

+++ 001.c
@@ -1,12 +1,12 @@
 #include <math.h>
 #include <stdio.h>
 
-int delta(int a, int b, int c) { return pow(b, 2) - 4 * a * c; }
+float delta(float a, float b, float c) { return pow(b, 2) - 4 * a * c; }
 
 int main() {
-  int a, b, c;
-  scanf("%d %d %d", &a, &b, &c);
-  printf("%d\n", delta(a, b, c));
+  float a, b, c;
+  scanf("%f %f %f", &a, &b, &c);
+  printf("%.2f\n", delta(a, b, c));
 
   return 0;
 }
